requirements.txt
asgiref==3.5.2
Django==3.2.4
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
django-phonenumber-field==6.3.0
phonenumberslite==8.12.52